+++ test.c
@@ -1,0 +1,6 @@
+#include <stdio.h>
+
+void main() {
+    printf("장우끼");
+    return 0;
+}
